test: move bria to correct dir
