+++ test/helpers/index.ts
@@ -20,6 +20,7 @@ export * from "./lightning"
 export * from "./price"
 export * from "./rate-limit"
 export * from "./redis"
+export * from "./shared"
 export * from "./state-setup-e2e"
 export * from "./user"
 export * from "./wallet"
