+++ test/helpers/lightning.ts
@@ -11,7 +11,6 @@ import {
 } from "@services/lnd/utils"
 import { baseLogger } from "@services/logger"
 import { WalletsRepository } from "@services/mongoose"
-import { NewOnChainService } from "@services/bria"
 
 import { sleep } from "@utils"
 
@@ -38,11 +37,11 @@ import { parsePaymentRequest } from "invoices"
 import {
   bitcoindClient,
   bitcoindOutside,
-  bitcoindSignerClient,
-  bitcoindSignerWallet,
   RANDOM_ADDRESS,
   sendToAddressAndConfirm,
 } from "./bitcoin-core"
+
+import { waitFor } from "./shared"
 
 export * from "lightning"
 
@@ -225,28 +224,6 @@ export const resetLnds = async () => {
   await mineBlockAndSync({ lnds })
 }
 
-export const resetBria = async () => {
-  const block = await bitcoindClient.getBlockCount()
-  if (!block) return // skip if we are just getting started
-
-  const existingSignerWallets = await bitcoindSignerClient.listWalletDir()
-  if (!existingSignerWallets.map((wallet) => wallet.name).includes("dev")) {
-    return
-  }
-
-  const balance = await bitcoindSignerWallet.getBalance()
-  if (balance === 0) return
-
-  await bitcoindSignerWallet.sendToAddress({
-    address: RANDOM_ADDRESS,
-    amount: balance,
-    subtractfeefromamount: true,
-  })
-  await bitcoindOutside.generateToAddress({ nblocks: 3, address: RANDOM_ADDRESS })
-
-  await waitUntilBriaZeroBalance()
-}
-
 export const closeAllChannels = async ({ lnd }) => {
   let channels
   try {
@@ -348,27 +325,6 @@ export const waitUntilChannelBalanceSync = ({ lnd }) =>
     const { unsettled_balance } = await getChannelBalance({ lnd })
     return unsettled_balance === 0
   })
-
-export const waitFor = async (f) => {
-  let res
-  while (!(res = await f())) await sleep(500)
-  return res
-}
-
-const waitUntilBriaZeroBalance = async () => {
-  await waitFor(async () => {
-    const balanceAmount = await NewOnChainService().getBalance()
-    if (balanceAmount instanceof Error) throw balanceAmount
-    const balance = Number(balanceAmount.amount)
-
-    if (balance > 0) {
-      baseLogger.warn({ briaBalance: `${balance} sats` }, "bria balance not zero yet")
-      return false
-    }
-
-    return true
-  })
-}
 
 export const waitUntilGraphIsReady = async ({ lnd, numNodes = 4 }) => {
   await waitFor(async () => {
