+++ test/helpers/shared.ts
@@ -1,0 +1,7 @@
+import { sleep } from "@utils"
+
+export const waitFor = async (f) => {
+  let res
+  while (!(res = await f())) await sleep(500)
+  return res
+}
